Quick note on using node version 8 (#29)
It would be helpful for folks to know (especially new developers) that they need to be running node version 8 in order to get the commands in this code-along to work properly. I am sure as you all know, node will run the latest version that is installed. I ran into errors running commands in this code-along before realizing I needed to switch versions. Anyways, keep up the great work!