Add a quick note to run ` nvm use 8 ` (#28)
It would be helpful to include a quick note (especially for new developers) to use the command ```nvm use 8``` to switch to the correct version of node so the composer can run correctly. As I am sure you all know, node by default uses the latest version and with node v10 some of the commands in this code-along won't execute correctly. Thanks and keep up the great work! Hope this helps.